+++ src/main/java/com/github/mjkuranda/spaceadventure2/SpaceAdventure2.java
@@ -1,5 +1,6 @@
 package com.github.mjkuranda.spaceadventure2;
 
+import com.github.mjkuranda.spaceadventure2.resources.GameImage;
 import com.github.mjkuranda.spaceadventure2.states.GameOverState;
 import com.github.mjkuranda.spaceadventure2.states.GameState;
 import com.github.mjkuranda.spaceadventure2.states.IntroState;
@@ -21,13 +22,13 @@ public class SpaceAdventure2 extends StateBasedGame {
 
     @Override
     public void initStatesList(GameContainer container) throws SlickException {
+        // FIXME: Loading resources in `LoadingState` state.
+        GameImage.load();
+
         addState(new MainMenuState(this));
         addState(new GameState());
         addState(new GameOverState());
         addState(new IntroState());
-
-        GameImages images = GameImages.getInstance();
-        images.init();
     }
 
     public static void main(String[] args) {