+++ src/main/java/com/github/mjkuranda/spaceadventure2/renderers/RetroRenderer.java
@@ -1,13 +1,10 @@
 package com.github.mjkuranda.spaceadventure2.renderers;
 
 import com.github.mjkuranda.spaceadventure2.GameData;
-import com.github.mjkuranda.spaceadventure2.GameImages;
 import com.github.mjkuranda.spaceadventure2.PlayerStatistics;
 import com.github.mjkuranda.spaceadventure2.entities.Entity;
-import com.github.mjkuranda.spaceadventure2.entities.EntityType;
 import com.github.mjkuranda.spaceadventure2.entities.Spaceship;
-import com.github.mjkuranda.spaceadventure2.entities.missiles.Missile;
-import com.github.mjkuranda.spaceadventure2.graphics.ImageFactory;
+import com.github.mjkuranda.spaceadventure2.resources.GameImage;
 import org.newdawn.slick.Color;
 import org.newdawn.slick.Graphics;
 import org.newdawn.slick.Image;
@@ -31,7 +28,7 @@ public class RetroRenderer extends Renderer {
         float playerX = data.getPlayer().getX();
         int xOffset = (int) ((playerX - 8.0f) * 8);
 
-        Image img = GameImages.BACKGROUND_IMAGE;
+        Image img = GameImage.BACKGROUND.get();
         float inc = PlayerStatistics.getInstance().getDistance() / 2;
         img.draw(-xOffset - 64 - (inc / 2), 0 - (inc / 2), img.getWidth() + inc, img.getHeight() + inc);
     }
@@ -81,7 +78,7 @@ public class RetroRenderer extends Renderer {
         int xPlayer = RENDERER_WIDTH / 2 - (playerWidth / 2);
         int yPlayer = RENDERER_HEIGHT - 218;
 
-        GameImages.SPACESHIP_IMAGE.draw(xPlayer, yPlayer, 0.37f);
+        GameImage.SPACESHIP.get().draw(xPlayer, yPlayer, 0.37f);
     }
 
     private void drawColumn(Graphics g, float x) {
@@ -139,7 +136,7 @@ public class RetroRenderer extends Renderer {
             return;
         }
 
-        Image img = ImageFactory.get(e.getType());
+        Image img = e.getImage();
         img.draw(x1 + xe - center, ye - center, size, size);
 
         //g.fillRect(x1 + xe - center, ye - center, size, size);