+++ src/main/java/com/github/mjkuranda/spaceadventure2/resources/GameResource.java
@@ -1,0 +1,11 @@
+package com.github.mjkuranda.spaceadventure2.resources;
+
+import org.newdawn.slick.SlickException;
+
+public interface GameResource<T> {
+    static void load() throws SlickException {}
+
+    T get();
+
+    String PREFIX_PATH = "src/main/resources/";
+}